# Server/Api/openapi-with-docs.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Dead Pidgeons API",
    "version": "v0.1"
  },
  "paths": {
...
    "/api/Balance/user/{userId}": {
      "get": {
        "tags": [
          "Balance"
        ],
        "operationId": "Balance_GetUserBalance",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Board/user/{userId}": {
      "get": {
        "tags": [
          "Board"
        ],
        "operationId": "Board_GetBoardsByUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardResponseDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "User_UpdateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updateDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDTO"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "operationId": "User_DeleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/create": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "User_CreateUser",
        "requestBody": {
          "x-name": "userDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDTO"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{id}/details": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetUserDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{id}/toggle-active": {
      "patch": {
        "tags": [
          "User"
        ],
        "operationId": "User_ToggleUserActiveStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{id}/set-active": {
      "patch": {
        "tags": [
          "User"
        ],
        "operationId": "User_SetUserActiveStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserActiveDTO"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{id}/set-admin": {
      "patch": {
        "tags": [
          "User"
        ],
        "operationId": "User_SetUserAdminStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserAdminDTO"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/search/{phonenumber}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_SearchUserByPhoneNumber",
        "parameters": [
          {
            "name": "phonenumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phonenumber": {
            "type": "string"
          },
          "isactive": {
            "type": "boolean"
          },
          "isadmin": {
            "type": "boolean"
          },
          "balance": {
            "type": "number",
            "format": "decimal"
          },
          "password": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "balancelogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Balancelog"
            }
          },
          "boards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          }
        }
      },
      "Balancelog": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userid": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "transactionnumber": {
            "type": "string"
          },
          "approved": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Board": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "userid": {
            "type": "string"
          },
          "selectednumbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "winner": {
            "type": "boolean"
          },
          "gameid": {
            "type": "string",
            "nullable": true
          },
          "game": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Game"
              }
            ]
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Game": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "weeknumber": {
            "type": "string"
          },
          "winningnumbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "drawdate": {
            "type": "string",
            "format": "date-time"
          },
          "isactive": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "boards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          }
        }
      },
...
      "BoardResponseDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "selectedNumbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "winner": {
            "type": "boolean"
          },
          "price": {
            "type": "number",
            "format": "decimal"
          }
        }
      },
...
      "CreateUserDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phonenumber": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "UpdateUserDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phonenumber": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SetUserActiveDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "SetUserAdminDTO": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isAdmin": {
            "type": "boolean"
          }
        }
      }
    }
  }
}